show beginning of last error message instead if the end
cb_appendInfoPopup sets the Text fields focus to the
end of the newly added message which together with the
Scroll code changing the start of a visible area to
the focused point even if there is nothing afterwards
causes sometimes no message being visible at all (as
reported in http://bugs.debian.org/597080).

Even disregarding the suboptimal behaviour of the Scroll
code, the end of a message is usually less interesting
than the beginning, so focus on the beginning of the last
written message instead.