avoid make distcheck failing because of gv-i18n.h
Make distcheck fails because a gv-i18n.h file is left over
after building and dist-cleaning.

To avoid this make gv-i18n.h not depend on the nls/ directory
(causing rebuilds of this file), but list the languages
to include in a nls/LANGUAGES files.

Additionally move it from EXTRA_DIST to gv_SOURCES proper.

Version 2 of this patch should fix some problems when building
directly from git and documents the replacement of .lnk files