do not define errno, as it might be a macro already.
Errno will usually be a macro nowadays (to make use of thread local
storage), so defining it as

extern int errno;

is made by the preprocessor to:

extern int (*__errno_location ());

which luckily currently still works, but shows that is really not
what should be done, so remove the definitions. (Which also silences
a compiler warning).