Revert "Fix Debian bug #594991: unexpected scrolling of the TOC widget"
This reverts commit 6568096a5e7fb72fb07780325e905c84d3c1003d.

Reverting fixes the following bug:

Using the END key (or whatever equivalent defined) does show the
last page, but the TOC is not scrolled to the number if the number
is outside the current TOC scroll window. It scrolls somewhere near
if the UP key is pressed afterwards. An annoying additional space
is shown beyond the last entry if you scroll down to the last position
using the slider.